+++ packages/tpetra/core/example/BlockCrs/CMakeLists.txt
@@ -3,7 +3,7 @@
 # It does not refer to MPI explicitly.
 TRIBITS_ADD_EXECUTABLE_AND_TEST(
   BlockCrsPerfTest
-  SOURCES driver.cpp
+  SOURCES Tpetra_TestBlockCrs.cpp
   ARGS 
   COMM  mpi
   STANDARD_PASS_OUTPUT
